+++ packages/sv/lib/addons/sveltekit-adapter/index.ts
@@ -65,9 +65,9 @@ export default defineAddon({
 			// in sk 3, we will keep "preview": "vite preview" like any other adapter
 			if (options.adapter === 'cloudflare') {
 				if (options.cfTarget === 'workers') {
-					data.scripts.preview = 'wrangler dev .svelte-kit/cloudflare/_worker.js';
+					data.scripts.preview = 'wrangler dev .svelte-kit/cloudflare/_worker.js --port 4173';
 				} else if (options.cfTarget === 'pages') {
-					data.scripts.preview = 'wrangler pages dev .svelte-kit/cloudflare';
+					data.scripts.preview = 'wrangler pages dev .svelte-kit/cloudflare --port 4173';
 				}
 			}
 
@@ -207,7 +207,7 @@ export default defineAddon({
 
 					data.compilerOptions ??= {};
 					data.compilerOptions.types ??= [];
-					data.compilerOptions.types.push('worker-configuration.d.ts');
+					data.compilerOptions.types.push('./worker-configuration.d.ts');
 
 					return generateCode();
 				});