+++ packages/sv/lib/create/index.ts
@@ -1,6 +1,7 @@
 import fs from 'node:fs';
 import path from 'node:path';
-import { mkdirp, copy, dist, getSharedFiles, replace } from './utils.ts';
+import { mkdirp, copy, dist, getSharedFiles } from './utils.ts';
+import { commonFilePaths } from '../cli/add/utils.ts';
 import { sanitizeName } from '../coreInternal.ts';
 
 export type TemplateType = (typeof templateTypes)[number];
@@ -82,8 +83,8 @@ function write_template_files(template: string, types: LanguageType, name: strin
 function write_common_files(cwd: string, options: Options, name: string) {
 	const files = getSharedFiles();
 
-	const pkg_file = path.join(cwd, 'package.json');
-	const pkg = JSON.parse(fs.readFileSync(pkg_file, 'utf-8'));
+	const pkg_file = path.join(cwd, commonFilePaths.packageJson);
+	const pkg = /** @type {any} */ JSON.parse(fs.readFileSync(pkg_file, 'utf-8'));
 
 	sort_files(files).forEach((file) => {
 		const include = file.include.every((condition) => matches_condition(condition, options));
@@ -91,7 +92,7 @@ function write_common_files(cwd: string, options: Options, name: string) {
 
 		if (exclude || !include) return;
 
-		if (file.name === 'package.json') {
+		if (file.name === commonFilePaths.packageJson) {
 			const new_pkg = JSON.parse(file.contents);
 			merge(pkg, new_pkg);
 		} else {